rag-customer-support/requirements.txt
langgraph
langchain
langchain-community
langchain-openai
chromadb
pypdf
tiktoken
python-dotenv
streamlit
fastapi
uvicorn
pydantic
